scripts/requirements.txt
PyYAML==3.11
oauthlib==0.7.2
requests==2.4.3
requests-oauthlib==0.4.1
six==1.7.3
tweepy==3.1.0
